+++ netbox_dns/templates/netbox_dns/dnssecpolicy.html
@@ -118,16 +118,6 @@
                         <th scope="row">{% trans "Parent Propagation Delay" %}</th>
                         <td>{{ object.parent_propagation_delay|placeholder }}</td>
                     </tr>
-                    <tr>
-                        <th scope="row">{% trans "Parental Agents" %}</th>
-                        <td>
-                            <table>
-                                {% for parental_agent in object.parental_agents %}
-                                <tr><td>{{ parental_agent }}</td></tr>
-                                {% endfor %}
-                            </table>
-                        </td>
-                    </tr>
                 </table>
             </div>
             <div class="card">
